requirements.txt
requests
python-telegram-bot==12.0.0b1
pyTelegramBotAPI==3.7.3

